Move predicates to pyobjects.py
