+++ rope/base/utils/predicates.py
@@ -2,30 +2,31 @@
 
 """
 
-import rope.base.builtins as builtins
-import rope.base.pyobjects as pyobjects
+
+# import rope.base.builtins as builtins
+# import rope.base.pyobjects as pyobjects
 
 
-def is_abstract_class(obj):
-    return isinstance(
-        obj,
-        (
-            builtins.BuiltinClass,
-            builtins.Generator,
-            builtins.Iterator,
-            pyobjects.PyClass,
-        ),
-    )
+# def is_abstract_class(obj):
+# return isinstance(
+# obj,
+# (
+# builtins.BuiltinClass,
+# builtins.Generator,
+# builtins.Iterator,
+# pyobjects.PyClass,
+# ),
+# )
 
 
-def is_abstract_function(obj):
-    return isinstance(
-        obj, (builtins.BuiltinFunction, builtins.Lambda, pyobjects.PyFunction)
-    )
+# def is_abstract_function(obj):
+# return isinstance(
+# obj, (builtins.BuiltinFunction, builtins.Lambda, pyobjects.PyFunction)
+# )
 
 
-def is_abstract_module(obj):
-    return isinstance(
-        obj,
-        (builtins.BuiltinModule, pyobjects._PyModule),
-    )
+# def is_abstract_module(obj):
+# return isinstance(
+# obj,
+# (builtins.BuiltinModule, pyobjects._PyModule),
+# )
