+++ rope/refactor/rename.py
@@ -12,7 +12,7 @@ from rope.base import (
 )
 from rope.base.change import ChangeContents, ChangeSet, MoveResource
 from rope.refactor import occurrences
-from rope.base.utils.predicates import is_abstract_module
+from rope.base.pyobjects import is_abstract_module
 
 
 class Rename:
